First commit in 3 years
First, changed the batch file so that the code is way more efficient (pursuant to issue #4 ). Secondly, updated the README. Finally, removed makerom.exe because I didn't make it, it's an outdated version (the latest release as of v0.18 works just fine), and it feels wrong to include it in this converter when you can just as easily download it.